# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the server side for MyDiary web app. MyDiary is an online journal where users can pen down their thoughts and feelings. \n You can find the app on [MyDiary](http://Alameen688.github.io/MyDiary/UI).",
    "version": "1.0.0",
    "title": "MyDiary API",
    "contact": {
      "email": "ogundiran12@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "mydiaryoxygen.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/entries": {
      "get": {
        "tags": [
          "entry"
        ],
        "summary": "get all entries by a user",
        "description": "This can only be done by the logged in user.",
        "operationId": "getAll",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EntryResponse"
              }
            }
          },
          "401": {
            "description": "You do not have access to this resource"
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      },
      "post": {
        "tags": [
          "entry"
        ],
        "summary": "add new entry",
        "description": "This can only be done by the logged in user.",
        "operationId": "create",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created entry object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EntryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/EntryResponse"
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "You do not have access to this resource"
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "name": "token",
      "in": "header"
    }
  },
  "definitions": {
    "EntryResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "description": "Response status",
          "enum": [
            "success",
            "error"
          ]
        },
        "data": {
          "$ref": "#/definitions/Entry"
        }
      },
      "xml": {
        "name": "EntryResponse"
      }
    },
    "Entry": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string",
          "example": "entry title"
        },
        "content": {
          "type": "string"
        },
        "user_id": {
          "type": "integer",
          "format": "int64"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      },
      "xml": {
        "name": "Entry"
      }
    },
    "EntryRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      },
      "xml": {
        "name": "EntryRequest"
      }
    },
...
  },
...
}